Explain how to find dediserver details for 3rd-party hosts
Each 3rd-party host does stuff differently, so there isn't a universal
"recipe" to follow to get the (S)FTP parameters to use with the mod
manager. However, most of this info is (should be) available in server
hosts' documentation on how to use (S)FTP, and finding the path is
relatively easy if one knows what to look for.