Update faq.adoc to include how to dual install a new and old game version, as well as removing uneccesary steps from the downgrading instructions
Hi,
As it says above, I tweaked the downgrading instructions to remove making a backup of the game files. Since the saves are stored seperatly, deleting the game files manually wont cause any data loss beyond the mod configs at worst. 

I also added instructions on how to have two installations of the game, one older version with mods and the steam installation on the latest update. Technically, this also counts as "how to pirate" if someone gives that copy to another person. HOWEVER, I would like to note that using a copied version you made yourself is not a crime, and that copy and pasting the game files is the first thing anyone would think of when attempting to pirate the game.

Please feel free to contact me with any questions, suggestions or anything at all!